New TeX mode
From an email on aspell-devel.
https://lists.gnu.org/archive/html/aspell-devel/2011-02/msg00000.html

A new TeX mode for Aspell 0.60.6
by Matthias Franz

This patch (hopefully) improves Aspell's TeX mode. The new features
are as follows:

1) Aspell skips over math contents inside $...$, $$..$$, \[...\] and
common LaTeX and AMS-LaTeX environments like "equation" and "gather".
Additional environments to be skipped over can be defined via the new
"tex-ignore-env" list.

2) Forced spell-checking of macro arguments while skipping over
arguments or environments. This is achieved with the new 'T' option to
"add-tex-command".

3) Discretionary hyphens and italic corrections are ignored. For
example, "hy\-phen" and "shelf\/ful" are recognized as single words.

4) Spell-checking can be manually switched on and off by putting the
text "aspell:on" and "aspell:off" into the file. This allows to skip
over macro definitions and other parts of text that confuse Aspell.

See the new Info file for more details.

Comments about the patch:

./modules/filter/context.cpp
It seems that the order of the context filter was undefined. Whether
one wants to have this filter before or after other filters depends on
the application in mind. I need it before the tex filter so that one
can place the "aspell:off" and "aspell:on" flags inside TeX comments.

./modules/filter/tex.cpp
The new code for the tex filter.

./modules/filter/tex-filter.info
The new default settings for the filter variables. "tex-ignore-env" is
new. Most changes to "tex-command" are related to the new 'T' option.
The new "begin" line is used internally. The rest are additions and
corrections to existing definitions.

./modules/filter/modes/tex.amf
Note that we enable the context filter by default.

./manual/aspell.1
./manual/aspell.texi
Updated documentation.