Use typo_edit_distance even when no keyboard def file is provided.
typo_edit_distance also accounts for difference the normal edit_distance
functions do not account for, such as the difference between "dog's"
and "dogs"

The results for common misspellings are actually _better_ when done this
way.  When typo_edit_distance is not used the results were noticeably
worse.

Also add regression test for when no keyboard def file is provided.