Don't allow UCS-2/4 encoding when set via a config file or the env.
This includes using a ConvKey rather than just a string when setting
the encoding.  The ConvKey contans an additional flag that will
indicate if UCS-2/4 is allowed.  To determine is UCS-2/4 is allowed a
new method, retrieve_value, is added to Config that returns additional
information on the value.  UCS-2/4 is only allowed if the Value object
returned from retrieve_value("encoding") has the secure flag set.