Enable maintainer mode when "autogen" is found in the root directory.
This effectively enables maintainer mode when the source is checked
out from git and disables it in release tarballs as the "autogen" file
is moved into the maintainer/ directory in the latter case.

Also remove "--enable-maintainer-mode" from the test build so that
when used in release tarballs it will test the source as is without
regenerating anything.